+++ models/prompt_templates.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 # pylint: disable=too-few-public-methods
-"""Sales Support Model (SSM) prompt templates"""
+"""Sales Support Model (hsr) prompt templates"""
 
 from langchain.prompts import PromptTemplate